Solve selection and heap sort conflict with header file

+++ 2-selection_sort.c
@@ -14,7 +14,7 @@ void selection_sort(int *array, size_t size)
 
 	for (curr_idx = 0; curr_idx < size - 1; curr_idx++)
 	{
-		min_idx = find_minimum_index(array, curr_idx, size);
+		min_idx = find_minimum_index(array, (int) curr_idx, size);
 
 		if (min_idx != curr_idx)
 		{
@@ -31,7 +31,7 @@ void selection_sort(int *array, size_t size)
  * @size: size of the array
  * Return: index of the minimum element
  */
-int find_minimum_index(const int *array, int start, size_t size)
+size_t find_minimum_index(const int *array, int start, size_t size)
 {
 	size_t j, min_idx = start;
